break interceptor out into a module
